+++ tests/python/pants_test/repo_scripts/BUILD
@@ -2,11 +2,9 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 python_tests(
-  name = 'test_git',
+  name = 'git_hooks',
   dependencies = [
     '3rdparty/python:future',
-    'src/python/pants/scm',
-    'src/python/pants/scm:git',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
     'src/python/pants/util:process_handler',
